Optimize batch reader

* add predicate filter for file and row groups

* add pixels predicate test

* replace tab with spaces

* basic cache framework

* fix nextbatch possible bugs

* temp sync

* add dynamic array

* add serialization and simple add

* add radix put and remove

* done with basic radix

* fix radix remove bugs

* replace hashmap inside node with fixed size array

* fix read batch logic

* optimize read batch performance, much improved

* refactor timestamp writer: each nano follows each time; add batch read opt for timestamp reader

* optimize string reader

* fix timestamp writer bug

* add string batch support

* basic tests passed for reader

Approved-by: Guodong Jin <guod.jin@gmail.com>

+++ pixels-core/src/test/java/cn/edu/ruc/iir/pixels/core/writer/TestPixelsWriter.java
@@ -52,7 +52,7 @@ public class TestPixelsWriter {
             PixelsWriter pixelsWriter =
                     PixelsWriterImpl.newBuilder()
                             .setSchema(schema)
-                            .setPixelStride((int) TestParams.pixelStride)
+                            .setPixelStride(TestParams.pixelStride)
                             .setRowGroupSize(TestParams.rowGroupSize)
                             .setFS(fs)
                             .setFilePath(new Path(filePath))
@@ -60,7 +60,7 @@ public class TestPixelsWriter {
                             .setReplication(TestParams.blockReplication)
                             .setBlockPadding(TestParams.blockPadding)
                             .setEncoding(TestParams.encoding)
-                            .setCompressionBlockSize((int) TestParams.compressionBlockSize)
+                            .setCompressionBlockSize(TestParams.compressionBlockSize)
                             .build();
 
             long curT = System.currentTimeMillis();
@@ -114,7 +114,7 @@ public class TestPixelsWriter {
             PixelsWriter pixelsWriter =
                     PixelsWriterImpl.newBuilder()
                             .setSchema(schema)
-                            .setPixelStride((int) TestParams.pixelStride)
+                            .setPixelStride(TestParams.pixelStride)
                             .setRowGroupSize(TestParams.rowGroupSize)
                             .setFS(fs)
                             .setFilePath(new Path(filePath))
@@ -122,7 +122,7 @@ public class TestPixelsWriter {
                             .setReplication(TestParams.blockReplication)
                             .setBlockPadding(TestParams.blockPadding)
                             .setEncoding(TestParams.encoding)
-                            .setCompressionBlockSize((int) TestParams.compressionBlockSize)
+                            .setCompressionBlockSize(TestParams.compressionBlockSize)
                             .build();
 
             long curT = System.currentTimeMillis();
